refactor: changed xtgeo_system to sys
