+++ src/xtgeo/__init__.py
@@ -92,7 +92,7 @@ from xtgeo.common.exceptions import GridNotFoundError
 from xtgeo.common.exceptions import BlockedWellsNotFoundError
 
 from xtgeo.common.xtgeo_dialog import XTGeoDialog
-from xtgeo.common.xtgeo_system import _XTGeoCFile
+from xtgeo.common.sys import _XTGeoCFile
 
 _xprint("Import common... done")
