+++ src/xtgeo/common/sys.py
