+++ src/xtgeo/grid3d/_gridprop_import_eclrun.py
@@ -8,7 +8,7 @@ import numpy.ma as ma
 import xtgeo
 import xtgeo.cxtgeo.cxtgeo as _cxtgeo
 
-from xtgeo.common import xtgeo_system as xsys
+from xtgeo.common import sys as xsys
 
 from . import _grid_eclbin_record as _eclbin
 from . import _grid3d_utils as utils
